+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -68,6 +68,7 @@ import org.graalvm.collections.EconomicSet;
 import org.graalvm.options.OptionDescriptors;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Engine;
+import org.graalvm.polyglot.EnvironmentAccess;
 import org.graalvm.polyglot.HostAccess;
 import org.graalvm.polyglot.Instrument;
 import org.graalvm.polyglot.Language;
@@ -81,6 +82,7 @@ import org.graalvm.polyglot.Value;
 import org.graalvm.polyglot.io.ByteSequence;
 import org.graalvm.polyglot.io.FileSystem;
 import org.graalvm.polyglot.io.MessageTransport;
+import org.graalvm.polyglot.io.ProcessHandler;
 import org.graalvm.polyglot.management.ExecutionEvent;
 
 @SuppressWarnings("unused")
@@ -100,6 +102,17 @@ public abstract class AbstractPolyglotImpl {
         }
 
         public abstract ExecutionEvent newExecutionEvent(Object event);
+    }
+
+    public abstract static class IOAccess {
+        protected IOAccess() {
+            if (!getClass().getCanonicalName().equals("org.graalvm.polyglot.io.ProcessHandler.ProcessCommand.IOAccessImpl")) {
+                throw new AssertionError("Only one implementation of IOAccess allowed. " + getClass().getCanonicalName());
+            }
+        }
+
+        public abstract ProcessHandler.ProcessCommand newProcessCommand(List<String> cmd, String cwd, Map<String, String> environment, boolean redirectErrorStream,
+                        ProcessHandler.Redirect inputRedirect, ProcessHandler.Redirect outputRedirect, ProcessHandler.Redirect errorRedirect);
     }
 
     public abstract static class APIAccess {
@@ -164,6 +177,7 @@ public abstract class AbstractPolyglotImpl {
 
     APIAccess api;
     MonitoringAccess monitoring;
+    IOAccess io;
 
     public final void setMonitoring(MonitoringAccess monitoring) {
         this.monitoring = monitoring;
@@ -174,12 +188,28 @@ public abstract class AbstractPolyglotImpl {
         initialize();
     }
 
+    public final void setIO(IOAccess ioAccess) {
+        Objects.requireNonNull(ioAccess, "IOAccess must be non null.");
+        this.io = ioAccess;
+    }
+
     public APIAccess getAPIAccess() {
         return api;
     }
 
     public MonitoringAccess getMonitoring() {
         return monitoring;
+    }
+
+    public final IOAccess getIO() {
+        if (io == null) {
+            try {
+                Class.forName(ProcessHandler.ProcessCommand.class.getName(), true, getClass().getClassLoader());
+            } catch (ClassNotFoundException e) {
+                throw new IllegalStateException(e);
+            }
+        }
+        return io;
     }
 
     protected void initialize() {
@@ -397,7 +427,8 @@ public abstract class AbstractPolyglotImpl {
                         PolyglotAccess polyglotAccess,
                         boolean allowNativeAccess, boolean allowCreateThread, boolean allowHostIO, boolean allowHostClassLoading, boolean allowExperimentalOptions, Predicate<String> classFilter,
                         Map<String, String> options,
-                        Map<String, String[]> arguments, String[] onlyLanguages, FileSystem fileSystem, Object logHandlerOrStream);
+                        Map<String, String[]> arguments, String[] onlyLanguages, FileSystem fileSystem, Object logHandlerOrStream, boolean allowCreateProcess, ProcessHandler processHandler,
+                        EnvironmentAccess environmentAccess, Map<String, String> environment);
 
         public abstract String getImplementationName();