+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotContextConfig.java
@@ -53,6 +53,9 @@ import org.graalvm.polyglot.PolyglotAccess;
 import org.graalvm.polyglot.io.FileSystem;
 
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
+import java.util.Collections;
+import org.graalvm.polyglot.EnvironmentAccess;
+import org.graalvm.polyglot.io.ProcessHandler;
 
 final class PolyglotContextConfig {
     private static final String[] EMPTY_STRING_ARRAY = new String[0];
@@ -64,6 +67,7 @@ final class PolyglotContextConfig {
     final boolean nativeAccessAllowed;
     final boolean createThreadAllowed;
     final boolean hostClassLoadingAllowed;
+    final boolean createProcessAllowed;
     final Predicate<String> classFilter;
     private final Map<String, String[]> applicationArguments;
     final Set<String> allowedPublicLanguages;
@@ -72,15 +76,21 @@ final class PolyglotContextConfig {
     final Map<String, Level> logLevels;    // effectively final
     final Handler logHandler;
     final PolyglotAccess polyglotAccess;
+    final ProcessHandler processHandler;
+    final EnvironmentAccess environmentAccess;
+    private final Map<String, String> environment;
+    private volatile Map<String, String> configuredEnvironement;
 
     PolyglotContextConfig(PolyglotEngineImpl engine, OutputStream out, OutputStream err, InputStream in,
                     boolean hostLookupAllowed, PolyglotAccess polyglotAccess, boolean nativeAccessAllowed, boolean createThreadAllowed,
                     boolean hostClassLoadingAllowed, boolean allowExperimentalOptions,
                     Predicate<String> classFilter, Map<String, String[]> applicationArguments,
-                    Set<String> allowedPublicLanguages, Map<String, String> options, FileSystem fileSystem, Handler logHandler) {
+                    Set<String> allowedPublicLanguages, Map<String, String> options, FileSystem fileSystem, Handler logHandler,
+                    boolean createProcessAllowed, ProcessHandler processHandler, EnvironmentAccess environmentAccess, Map<String, String> environment) {
         assert out != null;
         assert err != null;
         assert in != null;
+        assert environmentAccess != null;
         this.out = out;
         this.err = err;
         this.in = in;
@@ -89,6 +99,7 @@ final class PolyglotContextConfig {
         this.nativeAccessAllowed = nativeAccessAllowed;
         this.createThreadAllowed = createThreadAllowed;
         this.hostClassLoadingAllowed = hostClassLoadingAllowed;
+        this.createProcessAllowed = createProcessAllowed;
         this.classFilter = classFilter;
         this.applicationArguments = applicationArguments;
         this.allowedPublicLanguages = allowedPublicLanguages;
@@ -111,6 +122,9 @@ final class PolyglotContextConfig {
             }
             languageOptions.put(optionKey, options.get(optionKey), allowExperimentalOptions);
         }
+        this.processHandler = processHandler;
+        this.environmentAccess = environmentAccess;
+        this.environment = environment == null ? Collections.emptyMap() : environment;
     }
 
     boolean isAccessPermitted(PolyglotLanguage from, PolyglotLanguage to) {
@@ -155,6 +169,31 @@ final class PolyglotContextConfig {
             values = lang.getOptionValues();
         }
         return values.copy();
+    }
+
+    Map<String, String> getEnvironment() {
+        Map<String, String> result = configuredEnvironement;
+        if (result == null) {
+            synchronized (this) {
+                result = configuredEnvironement;
+                if (result == null) {
+                    if (environmentAccess == EnvironmentAccess.NONE) {
+                        result = Collections.unmodifiableMap(environment);
+                    } else if (environmentAccess == EnvironmentAccess.INHERIT) {
+                        result = System.getenv();  // System.getenv returns unmodifiable map.
+                        if (!environment.isEmpty()) {
+                            result = new HashMap<>(result);
+                            result.putAll(environment);
+                            result = Collections.unmodifiableMap(result);
+                        }
+                    } else {
+                        throw new IllegalStateException(String.format("Unsupported EnvironmentAccess: %s", environmentAccess));
+                    }
+                    configuredEnvironement = result;
+                }
+            }
+        }
+        return result;
     }
 
     private static PolyglotLanguage findLanguageForOption(PolyglotEngineImpl engine, final String optionKey, String group) {